backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
